gizmo_anki_plugin/requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0

# Optional - PDF import
pymupdf>=1.22.0; sys_platform != "win32"  # Mac/Linux

# Alternative PDF (poppler-utils CLI)
# Install via: brew install poppler (macOS) or apt install poppler-utils (Ubuntu)

# Optional - PowerPoint import
python-pptx>=0.6.21

# Optional - YouTube import
yt-dlp>=2023.0.0